Separa os arquivos .js nas pastas services e pages
