+++ assets/js/services/justification-service.js
@@ -1,6 +1,6 @@
-import { getUserToken } from './auth.js'
-import { backendUrl } from './backend.js'
-import handleErrors from './handle-errors.js'
+import { getUserToken } from '../auth.js'
+import { backendUrl } from '../backend.js'
+import handleErrors from '../handle-errors.js'
 
 export default class JustificationService {
     constructor() {
