+++ dashboard.html
@@ -66,7 +66,7 @@
         <!-- FIM CONTENT-CONTAINER -->
     </div>
     <!-- FIM MAIN-CONTAINER -->
-    <script type="module" src="assets/js/dashboard.js"></script>
+    <script type="module" src="assets/js/pages/dashboard.js"></script>
 </body>
 
 </html>
