+++ editar-horas.html
@@ -136,7 +136,7 @@
             </div>
         </div>
     </div>
-    <script type="module" src="assets/js/editar-horas.js"></script>
+    <script type="module" src="assets/js/pages/editar-horas.js"></script>
 </body>
 
 </html>
