+++ index.html
@@ -29,7 +29,7 @@
     </div>
     <!-- FIM MAIN-CONTAINER -->
 
-    <script src="assets/js/index.js"></script>
+    <script src="assets/js/pages/index.js"></script>
 </body>
 
 </html>
