+++ inserir-horas.html
@@ -152,7 +152,7 @@
             </div>
         </div>
     </div>
-    <script type="module" src="assets/js/inserir-horas.js"></script>
+    <script type="module" src="assets/js/pages/inserir-horas.js"></script>
 </body>
 
 </html>
