+++ login.html
@@ -90,7 +90,7 @@
 
     <!-- FIM MAIN-CONTAINER -->
 
-    <script type="module" src="assets/js/login.js"></script>
+    <script type="module" src="assets/js/pages/login.js"></script>
 </body>
 
 </html>
