+++ selecionar-perfil.html
@@ -50,7 +50,7 @@
         <!-- FIM SELECT PROFILE CONTAINER-->
     </div>
     <!-- FIM CONTENT CONTAINER-->
-    <script type="module" src="assets/js/selecionar-perfil.js"></script>
+    <script type="module" src="assets/js/pages/selecionar-perfil.js"></script>
 
 </body>
